Improved timestamp in filename
Now the timestamp in the filename actually is the timestamp of the picture, not when the picture was viewed.
The getTimestamp() method of the 'receivedSnap' instance is called for this. Unfortunately, this method belongs
to the superclass 'Snap', so I had to introduce some manual reflection instead of just using XposedHelpers.callMethod().